refactor(stories): better structure for stories
