--- examples/storybook-react/.babelrc.js
@@ -1,6 +1,0 @@
-const config = require('../../support/base.babel');
-
-module.exports = {
-  ...config,
-  sourceType: 'unambiguous',
-};
