--- examples/storybook-react/__e2e__/annotations.test.ts
@@ -1,30 +1,0 @@
-import { getDocument, queries } from 'playwright-testing-library';
-import { ElementHandle } from 'playwright-testing-library/dist/typedefs';
-import { goto, typist } from 'testing/playwright';
-
-const { getByRole, getByTitle } = queries;
-
-describe('Annotations', () => {
-  let $document: ElementHandle;
-  let $editor: ElementHandle;
-
-  describe('Basic', () => {
-    const path = 'annotation-extension--basic';
-
-    beforeEach(async () => {
-      await goto(path);
-      $document = await getDocument(page);
-      $editor = await getByRole($document, 'textbox');
-    });
-
-    it('should show the bubble menu', async () => {
-      await $editor.focus();
-      const $floatingMenu = await getByTitle($editor, 'Floating annotation');
-      await expect($floatingMenu.textContent()).resolves.toMatchInlineSnapshot(
-        `"This is a sample text"`,
-      );
-
-      await typist('{home} this should be at the end{enter}{enter}I am your fried!');
-    });
-  });
-});
