--- examples/storybook-react/__e2e__/tsconfig.json
@@ -1,67 +1,0 @@
-{
-  "__AUTO_GENERATED__": [
-    "To update the configuration edit the following field.",
-    "`package.json > @remirror > tsconfigs > '__e2e__'`",
-    "",
-    "Then run: `pnpm -w generate:ts`"
-  ],
-  "extends": "../../../support/tsconfig.base.json",
-  "compilerOptions": {
-    "types": [
-      "expect-playwright/global",
-      "jest-playwright-preset",
-      "jest",
-      "jest-extended",
-      "snapshot-diff",
-      "playwright",
-      "node"
-    ],
-    "declaration": false,
-    "noEmit": true,
-    "allowJs": true,
-    "skipLibCheck": true,
-    "importsNotUsedAsValues": "remove"
-  },
-  "include": ["./"],
-  "references": [
-    {
-      "path": "../../../packages/testing/src"
-    },
-    {
-      "path": "../../../packages/remirror/src"
-    },
-    {
-      "path": "../../../packages/remirror__core-helpers/src"
-    },
-    {
-      "path": "../../../packages/remirror__extension-codemirror5/src"
-    },
-    {
-      "path": "../../../packages/remirror__extension-heading/src"
-    },
-    {
-      "path": "../../../packages/remirror__extension-react-component/src"
-    },
-    {
-      "path": "../../../packages/remirror__extension-tables/src"
-    },
-    {
-      "path": "../../../packages/remirror__extension-text-color/src"
-    },
-    {
-      "path": "../../../packages/remirror__messages/src"
-    },
-    {
-      "path": "../../../packages/remirror__pm/src"
-    },
-    {
-      "path": "../../../packages/remirror__react/src"
-    },
-    {
-      "path": "../../../packages/create-context-state/src"
-    },
-    {
-      "path": "../../../packages/remirror/src"
-    }
-  ]
-}
