--- examples/storybook-react/package.json
@@ -1,69 +1,0 @@
-{
-  "name": "storybook-react",
-  "version": "0.0.0",
-  "private": true,
-  "description": "The storybook configuration for the remirror project.",
-  "homepage": "https://github.com/remirror/remirror/tree/HEAD/examples/storybook-react",
-  "repository": "https://github.com/remirror/remirror/tree/HEAD/examples/storybook-react",
-  "scripts": {
-    "build": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}'  build-storybook -c ./",
-    "e2e": "E2E_SERVER=storybook-react pnpm -w e2e",
-    "e2e:watch": "E2E_SERVER=storybook-react pnpm -w e2e:watch",
-    "start": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}'  start-storybook -p 6006 -c ./"
-  },
-  "dependencies": {
-    "@emotion/css": "^11.1.3",
-    "@linaria/core": "^3.0.0-beta.1",
-    "@remirror/core-helpers": "1.0.0-next.60",
-    "@remirror/extension-codemirror5": "1.0.0-next.60",
-    "@remirror/extension-heading": "1.0.0-next.60",
-    "@remirror/extension-react-component": "1.0.0-next.60",
-    "@remirror/extension-tables": "1.0.0-next.60",
-    "@remirror/extension-text-color": "0.0.0",
-    "@remirror/messages": "0.0.0",
-    "@remirror/pm": "1.0.0-next.60",
-    "@remirror/react": "1.0.0-next.60",
-    "@remirror/styles": "1.0.0-next.60",
-    "@storybook/addon-actions": "^6.1.17",
-    "@storybook/addon-links": "^6.1.17",
-    "@storybook/addons": "^6.1.17",
-    "@storybook/react": "^6.1.17",
-    "@types/codemirror": "^0.0.108",
-    "@types/refractor": "^3.0.0",
-    "clsx": "^1.1.1",
-    "codemirror": "^5.59.2",
-    "create-context-state": "0.0.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-is": "^17.0.1",
-    "reakit": "^1.3.5",
-    "refractor": "^3.3.1",
-    "remirror": "1.0.0-next.60"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.12.13",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.1",
-    "@types/webpack": "^4.41.26",
-    "@types/worker-plugin": "^4.0.0",
-    "babel-loader": "^8.2.2",
-    "monaco-editor-webpack-plugin": "^3.0.0",
-    "worker-plugin": "^5.0.0"
-  },
-  "@remirror": {
-    "tsconfigs": {
-      "src": {
-        "compilerOptions": {
-          "allowJs": true
-        }
-      },
-      "./": {
-        "include": [
-          "src",
-          ".babelrc.js",
-          "main.ts"
-        ]
-      }
-    }
-  }
-}
