--- examples/storybook-react/src/tsconfig.json
@@ -1,51 +1,0 @@
-{
-  "__AUTO_GENERATED__": [
-    "To update the configuration edit the following field.",
-    "`package.json > @remirror > tsconfigs > 'src'`",
-    "",
-    "Then run: `pnpm -w generate:ts`"
-  ],
-  "extends": "../../../support/tsconfig.base.json",
-  "compilerOptions": {
-    "types": [],
-    "declaration": false,
-    "noEmit": true,
-    "allowJs": true
-  },
-  "include": ["./"],
-  "references": [
-    {
-      "path": "../../../packages/remirror__core-helpers/src"
-    },
-    {
-      "path": "../../../packages/remirror__extension-codemirror5/src"
-    },
-    {
-      "path": "../../../packages/remirror__extension-heading/src"
-    },
-    {
-      "path": "../../../packages/remirror__extension-react-component/src"
-    },
-    {
-      "path": "../../../packages/remirror__extension-tables/src"
-    },
-    {
-      "path": "../../../packages/remirror__extension-text-color/src"
-    },
-    {
-      "path": "../../../packages/remirror__messages/src"
-    },
-    {
-      "path": "../../../packages/remirror__pm/src"
-    },
-    {
-      "path": "../../../packages/remirror__react/src"
-    },
-    {
-      "path": "../../../packages/create-context-state/src"
-    },
-    {
-      "path": "../../../packages/remirror/src"
-    }
-  ]
-}
