+++ examples/with-next/__e2e__/social-editor-ssr.test.ts
@@ -1,0 +1,7 @@
+import { smokeTest, ssrTest } from 'testing/playwright';
+
+const path = 'editor/social';
+
+smokeTest(path);
+ssrTest('social without content', path);
+ssrTest('social with content', 'editor/social/content');
