+++ examples/with-next/__e2e__/wysiwyg-editor-ssr.test.ts
@@ -1,0 +1,7 @@
+import { smokeTest, ssrTest } from 'testing/playwright';
+
+const path = 'editor/wysiwyg';
+
+smokeTest(path);
+ssrTest('wysiwyg without content', path);
+ssrTest('wysiwyg with content', 'editor/wysiwyg/content');
