+++ examples/with-next/package.json
@@ -14,7 +14,7 @@
     "@preconstruct/next": "^2.0.0",
     "@remirror/react": "1.0.0-next.60",
     "@remirror/styles": "1.0.0-next.60",
-    "next": "10.0.7",
+    "next": "10.0.8",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "refractor": "^3.3.1",
