--- examples/with-next/src/components/list-detail.tsx
@@ -1,16 +1,0 @@
-import { FunctionComponent } from 'react';
-
-import { User } from '../interfaces';
-
-interface ListDetailProps {
-  item: User;
-}
-
-const ListDetail: FunctionComponent<ListDetailProps> = ({ item: user }) => (
-  <div>
-    <h1>Detail for {user.name}</h1>
-    <p>ID: {user.id}</p>
-  </div>
-);
-
-export default ListDetail;
