--- examples/with-next/src/interfaces/index.ts
@@ -1,10 +1,0 @@
-// You can include shared interfaces/types in a separate file
-// and then use them in any component by importing them. For
-// example, to import the interface below do:
-//
-// import User from 'path/to/interfaces';
-
-export interface User {
-  id: number;
-  name: string;
-}
