+++ packages/storybook-react/.babelrc.js
@@ -1,0 +1,5 @@
+const config = require('../../support/base.babel');
+
+const plugins = [...config.plugins];
+
+module.exports = { ...config, plugins, sourceType: 'unambiguous' };
