+++ packages/storybook-react/main.js
@@ -1,0 +1,3 @@
+const { main } = require('./storybook-react-main');
+
+module.exports = main;
