+++ packages/storybook-react/package.json
@@ -1,0 +1,83 @@
+{
+  "name": "storybook-react",
+  "version": "0.0.0",
+  "private": true,
+  "description": "The storybook configuration for the remirror project.",
+  "homepage": "https://github.com/remirror/remirror/tree/HEAD/packages/storybook-react",
+  "repository": "https://github.com/remirror/remirror/tree/HEAD/packages/storybook-react",
+  "exports": {
+    ".": {
+      "require": "./dist/storybook-react.cjs.js"
+    },
+    "./package.json": "./package.json",
+    "./storybook-react-main": {
+      "require": "./storybook-react-main/dist/storybook-react-storybook-react-main.cjs.js"
+    },
+    "./types/*": "./dist/declarations/src/*.d.ts"
+  },
+  "main": "dist/storybook-react.cjs.js",
+  "files": [
+    "dist",
+    "storybook-react-main"
+  ],
+  "scripts": {
+    "build": "build-storybook -c ./",
+    "e2e": "E2E_SERVER=storybook-react pnpm -w e2e",
+    "e2e:watch": "E2E_SERVER=storybook-react pnpm -w e2e:watch",
+    "start": "start-storybook -p 6006 -c ./"
+  },
+  "dependencies": {
+    "@babel/runtime": "^7.13.10",
+    "@emotion/css": "^11.1.3",
+    "@linaria/core": "^3.0.0-beta.1",
+    "@remirror/messages": "0.0.0",
+    "@remirror/pm": "1.0.0-next.60",
+    "@remirror/react": "1.0.0-next.60",
+    "@remirror/styles": "1.0.0-next.60",
+    "@storybook/addons": "^6.1.20",
+    "@storybook/cli": "^6.1.20",
+    "@storybook/components": "^6.1.20",
+    "@storybook/core": "^6.1.20",
+    "@storybook/react": "^6.1.20",
+    "@storybook/theming": "^6.1.20",
+    "@types/codemirror": "^0.0.108",
+    "@types/react": "^17.0.2",
+    "@types/react-dom": "^17.0.1",
+    "@types/refractor": "^3",
+    "@types/webpack": "^4.41.26",
+    "babel-loader": "^8.2.2",
+    "codemirror": "^5.59.4",
+    "create-context-state": "0.0.0",
+    "globby": "^11.0.2",
+    "is-ci": "^3.0.0",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-is": "^17.0.1",
+    "reakit": "^1.3.5",
+    "refractor": "^3.3.1",
+    "remirror": "1.0.0-next.60",
+    "scripts": "0.0.0"
+  },
+  "@remirror": {
+    "tsconfigs": {
+      "src": {
+        "compilerOptions": {
+          "allowJs": true
+        }
+      },
+      "./": {
+        "include": [
+          "src",
+          ".babelrc.js",
+          "main.ts"
+        ]
+      }
+    }
+  },
+  "preconstruct": {
+    "entrypoints": [
+      "index.ts",
+      "storybook-react-main.ts"
+    ]
+  }
+}
