+++ packages/storybook-react/src/tsconfig.json
@@ -1,0 +1,33 @@
+{
+  "__AUTO_GENERATED__": [
+    "To update the configuration edit the following field.",
+    "`package.json > @remirror > tsconfigs > 'src'`",
+    "",
+    "Then run: `pnpm -w generate:ts`"
+  ],
+  "extends": "../../../support/tsconfig.base.json",
+  "compilerOptions": {
+    "types": [],
+    "declaration": false,
+    "noEmit": true,
+    "allowJs": true
+  },
+  "include": ["./"],
+  "references": [
+    {
+      "path": "../../remirror__messages/src"
+    },
+    {
+      "path": "../../remirror__pm/src"
+    },
+    {
+      "path": "../../remirror__react/src"
+    },
+    {
+      "path": "../../create-context-state/src"
+    },
+    {
+      "path": "../../remirror/src"
+    }
+  ]
+}
