+++ packages/storybook-react/src/typings.d.ts
@@ -1,0 +1,6 @@
+declare module 'is-ci' {
+  const isCi: boolean;
+  export = isCi;
+}
+
+declare module '*.svg';
