+++ packages/storybook-react/storybook-react-main/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "dist/storybook-react-storybook-react-main.cjs.js"
+}
