+++ packages/testing/package.json
@@ -40,7 +40,7 @@
     "react"
   ],
   "dependencies": {
-    "@babel/runtime": "^7.13.7",
+    "@babel/runtime": "^7.13.10",
     "@ffmpeg-installer/ffmpeg": "^1.0.20",
     "@react-spring/mock-raf": "^1.1.1",
     "@remirror/core": "1.0.0-next.60",
@@ -48,26 +48,26 @@
     "@remirror/preset-core": "1.0.0-next.60",
     "@remirror/react": "1.0.0-next.60",
     "@testing-library/react": "^11.2.5",
-    "@testing-library/react-hooks": "^5.0.3",
+    "@testing-library/react-hooks": "^5.1.0",
     "@types/fluent-ffmpeg": "^2.1.16",
     "@types/min-document": "^2.19.0",
-    "@types/react-test-renderer": "^17.0.0",
+    "@types/react-test-renderer": "^17.0.1",
     "@types/signal-exit": "^3.0.0",
     "@types/wait-on": "^5.2.0",
     "delay": "^5.0.0",
-    "expect-playwright": "^0.3.2",
+    "expect-playwright": "^0.3.3",
     "fluent-ffmpeg": "^2.1.2",
     "jest": "^26.6.3",
     "jest-dev-server": "^4.4.0",
     "jest-diff": "^26.6.2",
-    "jest-playwright-preset": "^1.4.6",
+    "jest-playwright-preset": "^1.5.1",
     "jest-process-manager": "^0.2.9",
     "looks-same": "^7.3.0",
     "min-document": "^2.19.0",
     "p-limit": "^3.1.0",
-    "playwright": "^1.8.0",
-    "playwright-core": "^1.8.0",
-    "playwright-testing-library": "^2.7.0",
+    "playwright": "^1.9.2",
+    "playwright-core": "^1.9.2",
+    "playwright-testing-library": "^2.7.2",
     "react-test-renderer": "^17.0.1",
     "signal-exit": "^3.0.3",
     "test-keyboard": "1.0.0-next.60"
